+++ R/zzz.R
@@ -1,3 +1,3 @@
 .onAttach <- function(...) {
-  packageStartupMessage("Welcome to Prevision.io R SDK [11.3.1]")
+  packageStartupMessage("Welcome to Prevision.io R SDK [11.3.3]")
 }