+++ NEWS.md
@@ -1,5 +1,23 @@
 # NEWS
 
+## 11.3.3
+
+Release date - 16th November 2021
+
+Cosmetic:
+  - Version bump to 11.3.3
+  - Minor documentation update
+  - Test function changed to remove deprecated calls
+
+Bug fix:
+  - Removed support of APP deployment
+  - Removed support of pipeline component creation
+  - Removed support of folder for data frame creation
+
+## 11.3.2
+
+No change in R SDK
+
 ## 11.3.1
 
 Release date - 25th October 2021